+++ pyqcrbox/pyqcrbox/registry/server/api_endpoints.py
@@ -64,10 +64,16 @@ async def retrieve_commands(
         (application_slug is None) or (sql_models_NEW_v2.ApplicationSpecDB.slug == application_slug),
         (application_version is None) or (sql_models_NEW_v2.ApplicationSpecDB.version == application_version),
     ]
+    stmt = select(
+        sql_models_NEW_v2.CommandSpecDB,
+        sql_models_NEW_v2.ApplicationSpecDB.slug,
+        sql_models_NEW_v2.ApplicationSpecDB.version).join(
+        sql_models_NEW_v2.CommandSpecDB.application,
+    )
 
     with settings.db.get_session() as session:
         # commands = session.scalars(select(model_cls).where(*filter_clauses)).all()
-        commands = session.scalars(select(model_cls)).all()
+        commands = session.scalars(stmt).all()
         commands = [cmd.to_response_model() for cmd in commands]
         return commands