+++ docs/CHANGELOG.md
@@ -19,6 +19,8 @@ However, there will be an initial period of stabilisation where this is not adhe
 - Added Traefik as a dynamic reverse proxy. ([#225](https://github.com/QCrBox/QCrBox/issues/225))
 - Added Devbox configuration to enable reproducible, automated setup. ([#258](https://github.com/QCrBox/QCrBox/issues/258))
 - Added script for a guided, automated installation of the QCrBox development environment.
+- Added 'qcb validate' command to convenient allow validation of application config yaml files.
+- Added support for 'one_of' in application config yaml files to specify alternative options for cif entries.
 
 ### Enhancements